refactor StateManager - Part 1
